+++ api/GraphQL/Query/UsersQuery.php
@@ -1,0 +1,72 @@
+<?php
+
+namespace Api\GraphQL\Query;
+
+use App\Model\User;
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Facades\GraphQL;
+use Rebing\GraphQL\Support\Query;
+use Rebing\GraphQL\Support\SelectFields;
+
+class UsersQuery extends Query
+{
+    protected $attributes = [
+        'name' => 'Users Query',
+        'description' => 'A query of users'
+    ];
+
+    public function type()
+    {
+        return GraphQL::paginate('users');
+    }
+
+    public function args()
+    {
+        return [
+            'id' => [
+                'name' => 'id',
+                'type' => Type::int()
+            ],
+            'name' => [
+                'name' => 'name',
+                'type' => Type::string()
+            ],
+            'email' => [
+                'name' => 'email',
+                'type' => Type::string()
+            ],
+            'photo_url' => [
+                'name' => 'photo_url',
+                'type' => Type::string()
+            ],
+            'page' => [
+                'name' => 'page',
+                'type' => Type::int(),
+                'description' => 'Display a specific page',
+            ],
+            'limit' => [
+                'name' => 'limit',
+                'type' => Type::int(),
+                'description' => 'Limit the items per page',
+            ],
+        ];
+    }
+
+    public function resolve($root, $args, SelectFields $fields)
+    {
+        $where = function ($query) use ($args) {
+            if (isset($args['id'])) {
+                $query->where('id',$args['id']);
+            }
+
+            if (isset($args['email'])) {
+                $query->where('email',$args['email']);
+            }
+        };
+        $user = User::with(array_keys($fields->getRelations()))
+            ->where($where)
+            ->select($fields->getSelect())
+            ->paginate($args['limit'] ?? 20, ['*'], 'page', $args['page'] ?? 0);
+        return $user;
+    }
+}