+++ api/GraphQL/Type/UserProfilesType.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Api\GraphQL\Type;
+
+use App\Model\Profile;
+use GraphQL\Type\Definition\Type;
+use Rebing\GraphQL\Support\Type as GraphQLType;
+
+class UserProfilesType extends GraphQLType
+{
+    protected $attributes = [
+        'name' => 'Profile',
+        'description' => 'A type',
+        'model' => Profile::class
+    ];
+
+    public function fields()
+    {
+        return [
+            'avatar' => [
+                'type' => Type::string(),
+                'description' => 'The avatar of user'
+            ],
+            'first_name' => [
+                'type' => Type::string(),
+                'description' => 'The first name of user'
+            ],
+            'last_name' => [
+                'type' => Type::string(),
+                'description' => 'The last name of user'
+            ]
+        ];
+    }
+}