+++ app/Http/Controllers/Settings/PasswordController.php
@@ -20,7 +20,7 @@ class PasswordController extends Controller
         ]);
 
         $request->user()->update([
-            'password' => bcrypt($request->password),
+            'password' => $request->password,
         ]);
     }
 }