+++ routes/web.php
@@ -11,6 +11,7 @@
 |
 */
 
-Route::get('{path}', function () {
-    return file_get_contents(public_path('_nuxt/index.html'));
-})->where('path', '(.*)');
+//! Uncomment if You are Using SPA Mode in Nuxt
+// Route::get('{path}', function () {
+//     return file_get_contents(public_path('_nuxt/index.html'));
+// })->where('path', '(.*)');