+++ tests/Feature/SettingsTest.php
@@ -2,13 +2,13 @@
 
 namespace Tests\Feature;
 
-use App\User;
+use App\Model\User;
 use Tests\TestCase;
 use Illuminate\Support\Facades\Hash;
 
 class SettingsTest extends TestCase
 {
-    /** @var \App\User */
+    /** @var \App\Model\User */
     protected $user;
 
     public function setUp()